refactor(panes): P17 #1 — stable pane IDs (eliminate the index-drift bug class)
PaneSlot was `usize` (raw Vec<Chart> index). Now `u64` (stable per-chart
ID allocated on first sight, never reused). The translation layer:

  Watchlist::pane_ids: Vec<u64>     // parallel to panes: Vec<Chart>
  Watchlist::next_pane_id: u64       // monotonic; 0 reserved as sentinel
  chart_idx_for(wl, id) -> Option<usize>   // id -> current index

Helpers added in gpu.rs to keep panes + pane_ids in lockstep:
- ensure_pane_ids_synced — lazy migration for older workspaces
- alloc_pane_id          — fresh id (no Vec mutation)
- push_chart_with_id     — push to both Vecs, return new id
- remove_chart_at        — remove from both, return dead id
- truncate_charts        — truncate both, return dropped ids

Pane mutations now flow through IDs end-to-end:
- Render loop: pl.iter_slots() → translate each id → current chart_idx,
  then index pane_rects scattered by chart_idx as before
- Close button: read id from pane_ids[idx] BEFORE Vec::remove, pass id
  to pl.close_pane(id)
- Close drain: remove_chart_at returns dead_id → pl.close_pane(dead_id)
- Split popup: push_chart_with_id returns new_id → pl.split_pane(src_id,
  axis, new_id). Source's id resolved from pane_ids[popup_pane_idx]
- Template regen (3 sites): build slot_ids slice from pane_ids[..max]
- Sync indicator: reference tree built with the SAME ids the live tree
  has (otherwise PartialEq would always fail on id mismatch)
- Defensive prune: leaf whose id ∉ pane_ids is closed (was: slot >= len)

P16 fix #2 "decrement every leaf slot > idx after panes.remove" is
DELETED — stable IDs make it impossible for one pane's close to
invalidate another's slot reference. ~15 lines removed from the close
drain, the entire correctness invariant now provable.

Tests: 14 pane_layout tests still pass (test slot literals updated to
u64; HashSet/Vec types updated; semantics unchanged).

+++ src-tauri/src/chart/renderer/pane_layout.rs
@@ -35,8 +35,12 @@ use crate::chart_renderer::gpu::Layout;
 use crate::ui_kit::widgets::pane_grid::{Axis, Node, PaneId, PaneState, SplitId, split_rect};
 use crate::ui_kit::widgets::theme::ComponentTheme;
 
-/// Leaf content — index into the chart-app's `Vec<Chart>` storage.
-pub type PaneSlot = usize;
+/// Leaf content — a stable u64 id assigned by `Watchlist::alloc_pane_id` and
+/// stored in lockstep with the chart in `Watchlist::pane_ids`. P17 #1 swap
+/// from raw `usize` indices to stable IDs eliminates the entire class of
+/// "leaf shifted to wrong chart after a close" bugs that previously required
+/// manual slot-decrement fixups in the close drain.
+pub type PaneSlot = u64;
 
 /// Recursive split-pane layout for the chart area. Owns the topology +
 /// split ratios; chart data lives elsewhere indexed by `PaneSlot`.
@@ -69,7 +73,7 @@ impl PartialEq for PaneLayout {
 
 impl Default for PaneLayout {
     fn default() -> Self {
-        let (state, _root) = PaneState::new(0_usize);
+        let (state, _root) = PaneState::new(0_u64);
         Self { state }
     }
 }
@@ -486,7 +490,7 @@ mod tests {
         assert!(l.split_pane(0, Axis::Vertical, 2).is_some());
         assert_eq!(l.pane_count(), 2);
         // slots reachable via iteration are exactly {0, 2}
-        let slots: std::collections::HashSet<usize> = l.iter_slots().map(|(_, s)| s).collect();
+        let slots: std::collections::HashSet<u64> = l.iter_slots().map(|(_, s)| s).collect();
         assert_eq!(slots, [0, 2].into_iter().collect());
     }
 
@@ -496,7 +500,7 @@ mod tests {
         assert_eq!(l.pane_count(), 2);
         l.replace_with_template(Layout::Four, &[0, 1, 2, 3]);
         assert_eq!(l.pane_count(), 4);
-        let slots: std::collections::HashSet<usize> = l.iter_slots().map(|(_, s)| s).collect();
+        let slots: std::collections::HashSet<u64> = l.iter_slots().map(|(_, s)| s).collect();
         assert_eq!(slots, [0, 1, 2, 3].into_iter().collect());
     }
 
@@ -530,8 +534,8 @@ mod tests {
         // Serialize → deserialize via the Clone impl (which round-trips JSON).
         let cloned = l.clone();
         assert_eq!(cloned.pane_count(), 2);
-        let cloned_slots: Vec<usize> = cloned.iter_slots().map(|(_, s)| s).collect();
-        let orig_slots:   Vec<usize> = l.iter_slots().map(|(_, s)| s).collect();
+        let cloned_slots: Vec<u64> = cloned.iter_slots().map(|(_, s)| s).collect();
+        let orig_slots:   Vec<u64> = l.iter_slots().map(|(_, s)| s).collect();
         assert_eq!(cloned_slots, orig_slots);
     }
